Split types
